Rename `delta` to `delta_lake`
